Centralize mapNode logic in one file
