+++ packages/dds/tree/src/simple-tree/rawNode.ts
@@ -5,27 +5,20 @@
 
 import { Anchor, AnchorNode, FieldKey, TreeNodeSchemaIdentifier } from "../core/index.js";
 import {
-	FlexMapNodeSchema,
 	FlexTreeContext,
 	FlexTreeEntityKind,
 	FlexTreeField,
-	FlexTreeMapNode,
 	FlexTreeNode,
 	FlexTreeNodeEvents,
 	FlexTreeNodeSchema,
-	FlexTreeTypedField,
 	FlexTreeTypedNode,
-	FlexTreeUnboxField,
-	FlexibleFieldContent,
 	TreeStatus,
 	flexTreeMarker,
 } from "../feature-libraries/index.js";
 import { fail } from "../util/index.js";
 
-import { InsertableContent } from "./proxies.js";
-
 /** Stores the content of the raw node, i.e. the data that was passed to the factory */
-const nodeContent = Symbol();
+export const nodeContent = Symbol();
 interface HasNodeContent<T> {
 	[nodeContent]: T;
 }
@@ -47,25 +40,12 @@ export function extractRawNodeContent(node: object): object | undefined {
 }
 
 /**
- * Creates a node that pretends to satisfy the given schema while wrapping the given node content.
+ * Node that pretends to satisfy the given schema while wrapping the given node content.
  * Retrieve the node content via {@link extractRawNodeContent}.
  *
  * @remarks This is useful for creating "raw" nodes: nodes which capture data about a pending insertion but are not yet inserted.
  * These raw nodes can be then used on the right-hand side of an assignment (via `=`) to the tree.
  * However, many of their properties and methods are currently unimplemented and will error if accessed.
- */
-export function createRawNode(
-	schema: FlexTreeNodeSchema,
-	content: InsertableContent,
-): RawTreeNode<FlexTreeNodeSchema, InsertableContent> {
-	if (schema instanceof FlexMapNodeSchema) {
-		return new RawMapNode(schema, content as ReadonlyMap<string, InsertableContent>);
-	}
-	fail("Unrecognized schema");
-}
-
-/**
- * The base implementation of a node created by {@link createRawNode}.
  */
 export abstract class RawTreeNode<TSchema extends FlexTreeNodeSchema, TContent>
 	implements FlexTreeNode, HasNodeContent<TContent>
@@ -76,6 +56,7 @@ export abstract class RawTreeNode<TSchema extends FlexTreeNodeSchema, TContent>
 	#anchor: Anchor | undefined;
 
 	public readonly type: TreeNodeSchemaIdentifier;
+
 	public constructor(
 		public readonly schema: TSchema,
 		content: TContent,
@@ -126,67 +107,6 @@ export abstract class RawTreeNode<TSchema extends FlexTreeNodeSchema, TContent>
 
 	public get anchorNode(): AnchorNode {
 		throw rawError("Reading anchor node");
-	}
-}
-
-/**
- * The implementation of a map node created by {@link createRawNode}.
- */
-export class RawMapNode<TSchema extends FlexMapNodeSchema>
-	extends RawTreeNode<TSchema, ReadonlyMap<string, InsertableContent>>
-	implements FlexTreeMapNode<TSchema>
-{
-	public get size(): number {
-		return this[nodeContent].size;
-	}
-	public has(key: string): boolean {
-		return this[nodeContent].has(key);
-	}
-	public get(key: string): FlexTreeUnboxField<TSchema["info"]> {
-		return this[nodeContent].get(key) as FlexTreeUnboxField<TSchema["info"]>;
-	}
-	public keys(): IterableIterator<FieldKey> {
-		return this[nodeContent].keys() as IterableIterator<FieldKey>;
-	}
-	public values(): IterableIterator<FlexTreeUnboxField<TSchema["info"], "notEmpty">> {
-		throw rawError("Iterating map values");
-	}
-	public entries(): IterableIterator<
-		[FieldKey, FlexTreeUnboxField<TSchema["info"], "notEmpty">]
-	> {
-		throw rawError("Iterating map entries");
-	}
-	public forEach(
-		callbackFn: (
-			value: FlexTreeUnboxField<TSchema["info"], "notEmpty">,
-			key: FieldKey,
-			map: FlexTreeMapNode<TSchema>,
-		) => void,
-		thisArg?: any,
-	): void {
-		throw rawError("Iterating maps with forEach");
-	}
-	public set(key: string, value: FlexibleFieldContent<TSchema["info"]> | undefined): void {
-		throw rawError("Setting a map entry");
-	}
-	public delete(key: string): void {
-		throw rawError("Deleting a map entry");
-	}
-
-	public get asObject(): {
-		readonly [P in FieldKey]?: FlexTreeUnboxField<TSchema["info"], "notEmpty">;
-	} {
-		throw rawError("Converting a map to an object");
-	}
-
-	public [Symbol.iterator](): IterableIterator<
-		[FieldKey, FlexTreeUnboxField<TSchema["info"], "notEmpty">]
-	> {
-		return this.entries();
-	}
-
-	public override boxedIterator(): IterableIterator<FlexTreeTypedField<TSchema["info"]>> {
-		throw rawError("Boxed iteration");
 	}
 }
